@@ -11,10 +11,10 @@ description: A new Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 0.0.2+2
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,11 +22,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  flutter_svg: 0.14.4
-  intl: ^0.15.8
-  flutter_bloc: ^0.19.0
-  equatable: ^0.2.6
+  cupertino_icons: ^0.1.3
+  flutter_svg: 0.17.1
+  intl: ^0.16.1
+  flutter_bloc: ^3.1.0
+  equatable: ^1.1.0
 
 dev_dependencies:
   flutter_test:
