@@ -5,13 +5,18 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 
-class Category extends Equatable{
+class Category extends Equatable {
   final int id;
   final String title;
   final String image;
   final int typeId;
 
-  Category({@required this.id, @required this.title, @required this.image,
-    @required this.typeId});
-  
+  Category(
+      {@required this.id,
+      @required this.title,
+      @required this.image,
+      @required this.typeId});
+
+  @override
+  List<Object> get props => [id, title, image, typeId];
 }
