@@ -5,7 +5,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 
-class Product extends Equatable{
+class Product extends Equatable {
   final int id;
   final String title;
   final String image;
@@ -17,8 +17,29 @@ class Product extends Equatable{
   final bool favorite;
   final bool isNew;
 
-  Product({@required this.id, @required this.title, @required this.image, 
-    @required this.categoryTitle, @required this.price, @required this.discountPercent, 
-    @required this.rating, @required this.ratingCount, @required this.favorite,
-    @required this.isNew});
+  Product(
+      {@required this.id,
+      @required this.title,
+      @required this.image,
+      @required this.categoryTitle,
+      @required this.price,
+      @required this.discountPercent,
+      @required this.rating,
+      @required this.ratingCount,
+      @required this.favorite,
+      @required this.isNew});
+
+  @override
+  List<Object> get props => [
+        id,
+        title,
+        image,
+        categoryTitle,
+        price,
+        discountPercent,
+        rating,
+        ratingCount,
+        favorite,
+        isNew
+      ];
 }
