@@ -17,12 +17,12 @@ class ProductBloc extends Bloc<ProductEvent, ProductState> {
   final ProductRepository productRepository;
   final CategoryRepository categoryRepository;
   final HashtagRepository hashtagRepository;
-  
-  ProductBloc(
-    {@required this.categoryRepository, 
+
+  ProductBloc({
+    @required this.categoryRepository,
     @required this.productRepository,
     @required this.hashtagRepository,
-  })  : assert(productRepository != null);
+  }) : assert(productRepository != null);
 
   @override
   ProductState get initialState => ProductInitialState();
@@ -30,40 +30,34 @@ class ProductBloc extends Bloc<ProductEvent, ProductState> {
   @override
   Stream<ProductState> mapEventToState(ProductEvent event) async* {
     ProductStateData data = ProductStateData();
-    if(event is ProductShowListEvent){
-      if ( this.currentState is ProductsListViewState) {
-        ProductsListViewState state = this.currentState as ProductsListViewState;
-        if ( state.category.id != event.categoryId ) {
+    if (event is ProductShowListEvent) {
+      if (this.state is ProductsListViewState) {
+        ProductsListViewState state = this.state as ProductsListViewState;
+        if (state.category.id != event.categoryId) {
           //Set state to loading
-          yield state.copyWith(
-            loading: true
-          );
+          yield state.copyWith(loading: true);
           //Load data from repositories
           data = getStateData(event.categoryId);
           //set state to loaded and update data
           yield state.copyWith(
-            products: data.products,
-            hashtags: data.hashtags,
-            loading: false,
-            category: data.category
-          );
-        } 
+              products: data.products,
+              hashtags: data.hashtags,
+              loading: false,
+              category: data.category);
+        }
       } else {
         data = getStateData(event.categoryId);
         yield ProductsListViewState(
-          isLoading: false,
-          category: data.category,
-          hashtags: data.hashtags,
-          products: data.products
-        );
+            isLoading: false,
+            category: data.category,
+            hashtags: data.hashtags,
+            products: data.products);
       }
-    } else if(event is ProductShowCardEvent){
-      if ( this.currentState is ProductsCardViewState) {
-        ProductsCardViewState state = this.currentState as ProductsCardViewState;
-        if ( state.category.id != event.categoryId ) {
-          yield state.copyWith(
-            loading: true
-          );
+    } else if (event is ProductShowCardEvent) {
+      if (this.state is ProductsCardViewState) {
+        ProductsCardViewState state = this.state as ProductsCardViewState;
+        if (state.category.id != event.categoryId) {
+          yield state.copyWith(loading: true);
           data = getStateData(event.categoryId);
           yield state.copyWith(
             products: data.products,
@@ -71,20 +65,19 @@ class ProductBloc extends Bloc<ProductEvent, ProductState> {
             category: data.category,
             hashtags: data.hashtags,
           );
-        } 
+        }
       } else {
         data = getStateData(event.categoryId);
         yield ProductsCardViewState(
-          isLoading: false,
-          category: data.category,
-          hashtags: data.hashtags,
-          products: data.products
-        );
+            isLoading: false,
+            category: data.category,
+            hashtags: data.hashtags,
+            products: data.products);
       }
     }
   }
-  
-  ProductStateData getStateData(int categoryId){
+
+  ProductStateData getStateData(int categoryId) {
     ProductStateData data = ProductStateData();
 
     data.products = productRepository.getProducts(categoryId);
@@ -93,10 +86,9 @@ class ProductBloc extends Bloc<ProductEvent, ProductState> {
 
     return data;
   }
-
 }
 
-class ProductStateData{
+class ProductStateData {
   List<Product> products;
   List<HashTag> hashtags;
   Category category;
