@@ -3,9 +3,10 @@
 // Date: 2020-02-06
 
 class OpenFlutterEcommerceRoutes {
-  static final home = "/";
+  static final home = "home";
   static final shop = "shop";
   static final cart = "cart";
   static final favourites = "favourites";
   static final profile = "profile";
+  static const SIGNUP = "signup";
 }
