@@ -7,25 +7,24 @@ import 'package:flutter/material.dart';
 @immutable
 abstract class CategoryEvent extends Equatable {
   final int typeId;
-  CategoryEvent(this.typeId, [List props = const []]) : super(props);
+  CategoryEvent(this.typeId) : super();
+
+  @override
+  List<Object> get props => [typeId];
 }
 
 @immutable
 class CategoryShowListEvent extends CategoryEvent {
-
   CategoryShowListEvent(int typeId) : super(typeId);
 
   @override
-  String toString() =>
-      'Category show list view';
+  String toString() => 'Category show list view';
 }
 
 @immutable
 class CategoryShowTilesEvent extends CategoryEvent {
-
   CategoryShowTilesEvent(typeId) : super(typeId);
 
   @override
-  String toString() =>
-      'Category show tiles view';
+  String toString() => 'Category show tiles view';
 }
