@@ -10,6 +10,8 @@ import 'package:openflutterecommerce/screens/home/home.dart';
 import 'package:openflutterecommerce/screens/wrapper.dart';
 import 'package:openflutterecommerce/widgets/scaffold.dart';
 
+import 'home.dart';
+
 class HomeScreen extends StatefulWidget {
   @override
   _HomeScreenState createState() => _HomeScreenState();
@@ -19,20 +21,17 @@ class _HomeScreenState extends State<HomeScreen> {
   @override
   Widget build(BuildContext context) {
     return SafeArea(
-      child: OpenFlutterScaffold(
-        background: null,
-        title: null,
-        body: BlocProvider<HomeBloc>(
-          builder: (context) {
-            return HomeBloc(
-              productRepository: ProductRepository()
-            )..dispatch(HomeLoadEvent());
+        child: OpenFlutterScaffold(
+      background: null,
+      title: null,
+      body: BlocProvider<HomeBloc>(
+          create: (context) {
+            return HomeBloc(productRepository: ProductRepository())
+              ..add(HomeLoadEvent());
           },
-          child: HomeWrapper()
-        ),
-        bottomMenuIndex: 0,
-      )
-    );
+          child: HomeWrapper()),
+      bottomMenuIndex: 0,
+    ));
   }
 }
 
@@ -42,27 +41,26 @@ class HomeWrapper extends StatefulWidget {
 }
 
 class _HomeWrapperState extends OpenFlutterWrapperState<HomeWrapper> {
-
   //State createState() => OpenFlutterWrapperState();
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<HomeEvent, HomeState>(
-        bloc: BlocProvider.of<HomeBloc>(context),
+    return BlocBuilder<HomeBloc, HomeState>(
         builder: (BuildContext context, HomeState state) {
-          return getPageView(<Widget>[
-            Main1View(changeView: changePage, 
-              products: state is HomeLoadedState?
-                state.newProducts : List<Product>()),
-            Main2View(changeView: changePage, 
-              salesProducts: state is HomeLoadedState?
-                state.salesProducts : List<Product>(), 
-              newProducts: state is HomeLoadedState?
-                state.newProducts : List<Product>()),
-            Main3View(changeView: changePage)
-          ]
-        );
-      }
-    );
+      return getPageView(<Widget>[
+        Main1View(
+            changeView: changePage,
+            products:
+                state is HomeLoadedState ? state.newProducts : List<Product>()),
+        Main2View(
+            changeView: changePage,
+            salesProducts: state is HomeLoadedState
+                ? state.salesProducts
+                : List<Product>(),
+            newProducts:
+                state is HomeLoadedState ? state.newProducts : List<Product>()),
+        Main3View(changeView: changePage)
+      ]);
+    });
   }
 }
